Fix for Issue #98 - BatchIf race condition
When the batch is resumed from a pause, the buffered changes are added
to a new ChangeSet. The changeset new's an Items list, but then assigns
the buffered list to the Items property.

This changeset is delivered to the consumer, then the buffer is cleared.
If the consumer is observing on a different thread, there is a race
between the changes being delivered and the buffer collection
(referenced by teh changeset) being cleared.

The fix here is to copy the changes from the buffer to the new'ed Items
list in the changeset.

Performance hit, but the race is eradicate and the new'ed Items list is
made use of.