Changes to allow compilation on VS 7.5+ for Mac
Reordering to put netstandard2.0 at the front of the target frameowkr list. This allows compilation and test execution to succeed on Visual Studio for ac (v7.5+) as VS for Mac picks the first target framework in the list to target currently. This fools it into picking netstandard2.0.

Also adjusted .gitignore to remove annoying OSX artifacts.