Rebuilt test fixtures for both `.ToObservableChangeSet()` operators, increasting code coverage from 69% to 100% across the internal implementation classes.
Identified existing defects within the `.ToObservableChangeSet()` operators, mostly related to error and completion propagation, and implemented testing for this behavior, with the tests being marked as "Skip" until the defects are fixed. This also includes the defect referred in #635.

Added an `IsCompleted` flag to `ChangeSetAggregator<T>`, to allow for testing of completion propagation.